+++ ChocAn.DataCenterConsole/Controllers/ProviderServiceController.cs
@@ -78,7 +78,7 @@ namespace ChocAn.DataCenterConsole.Controllers
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<IActionResult> DetailsAsync(decimal id)
+        public async Task<IActionResult> DetailsAsync(int id)
         {
             var action = new DetailsAction<Controller, ProviderService, ProviderServiceDetailsViewModel>(this, id, repository, mapper);
             return await action.ActionResult();
@@ -128,7 +128,7 @@ namespace ChocAn.DataCenterConsole.Controllers
         // POST: ProviderServiceController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Delete(decimal id)
+        public async Task<ActionResult> Delete(int id)
         {
             await repository.DeleteAsync(id);
             return RedirectToAction(nameof(Index));
