+++ ChocAn.ProviderService/Provider.cs
@@ -41,12 +41,12 @@ namespace ChocAn.ProviderRepository
     public class Provider
     {
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public decimal Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
         public string StreetAddress { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public decimal ZipCode { get; set; }
+        public int ZipCode { get; set; }
     }
 }
